% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-functions.R
\name{alphaq}
\alias{alphaq}
\title{Returns a function to retrieve the \eqn{\alpha}-quantile from a vector.}
\usage{
alphaq(a)
}
\arguments{
\item{a}{the quantile level \eqn{\alpha}}
}
\value{
Returns a function that takes an argument \code{x} and calculates the
        \eqn{\alpha}-quantile from the vector \code{x}.
}
\description{
To be used in
conjunction with apply.
}
\examples{

##############################################################################
}
