% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-functions.R
\name{arg.names}
\alias{arg.names}
\title{Returns the argument names of a function.}
\usage{
arg.names(fun)
}
\arguments{
\item{fun}{a function from which the arguments should be retrieved from}
}
\value{
a vector of characters containing the names of the arguments of the function \code{fun} 

##############################################################################
}
\description{
To check if an estimate or simulate function is valid.
}

