% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-functions.R
\name{compare.arg.names}
\alias{compare.arg.names}
\title{Compare arguments with character vector}
\usage{
compare.arg.names(fun, names)
}
\arguments{
\item{fun}{a function from which the arguments should be retrieved from}

\item{names}{names the arguments should be compared to}
}
\value{
Logical if the names of the arguments are identical to \code{names}

##############################################################################
}
\description{
To check if an estimate or simulate function is valid.
}

