% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QPBoot-class.R
\name{computeCIs-QPBoot}
\alias{computeCIs}
\alias{computeCIs,QPBoot-method}
\alias{computeCIs-QPBoot}
\title{Pointwise Confidence Intervalls}
\usage{
computeCIs(object, alpha = 0.05, method = c("quantiles", "norm"),
  levels = object@sPG@levels[[1]])
}
\arguments{
\item{object}{}

\item{alpha}{}

\item{method}{}
}
\value{
Returns a \code{list} with four elements
\item{\code{q_up}}{}
\item{\code{q_low}}{}
\item{\code{mean}}{}
\item{\code{sd}}{}
}
\description{
Depending on \code{method} this calculates pointwise confidence intervalls for
a smoothed periodgram that belongs to a time series defined by \code{model} and
\code{param}. If \code{(method = "quantiles")} it computes the \eqn{\alpha/2}
and \eqn{1-\alpha/2} quantiles from the Values of the simulated smoothed periodograms and
returns those. If \code{(method = "norm")} it uses the asymptotic normality
of the smoothed periodograms by estimating \code{mean} and
\code{standard deviation} for each frequency and computing the \eqn{\alpha/2}
and \eqn{1-\alpha/2} quantiles from a normal distribution with the estimated
parameters.
}

