% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dax}
\alias{dax}
\title{DAX: Deutscher Aktien Index 2000--2010}
\format{A univariate time series with 2802 observations; a \code{numeric} vector}
\description{
Contains the closing values of the DAX stock index for the years 2000--2010.
}
\details{
The data was downloaded from the YFinance section on the Quandl Website.
}
\examples{
plot(dax,type = "l")
}
\references{
Quandl
\url{https://www.quandl.com/data/YAHOO/INDEX_GDAXI-DAX-Index-Germany}
}
\keyword{data}

