% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsModel-class.R
\docType{methods}
\name{estimate-tsModel}
\alias{estimate,tsModel-method}
\alias{estimate-tsModel}
\title{Estimates the parameter of a \link{tsModel-class}}
\usage{
\S4method{estimate}{tsModel}(object, data)
}
\arguments{
\item{object}{the \link{tsModel-class} for that the parameters shall be estimated}

\item{data}{a univariate numeric vector containing the time-series data}
}
\value{
Returns the estimated parameters \code{par} and sets the \code{par} slot of the \link{tsModel-class}.
}
\description{
The methods estimates the parameters from the time-series in \code{data}, using the method
defined in the \link{tsModel-class}.
}

