% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{generics-accessors}
\alias{generics-accessors}
\alias{getCIs}
\title{Generic functions for accessing attributes of objects}
\usage{
getCIs(object, ...)
}
\arguments{
\item{object}{object from which to get the value}

\item{...}{optional parameters; for documentation see the documentation of
            the methods to each of the generic.}
}
\description{
These generic functions are needed to access the objects' attributes.
Note that the naming convention \code{getAttribute} was applied, where
\code{attribute} is the name of the attribute/slot of the class of the
object.
}

