% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{getAR}
\alias{ARp}
\alias{getAR}
\title{Autoregressive time series of order p}
\usage{
getAR(p)
}
\arguments{
\item{p}{numeric sets the order of the AR(p) process}
}
\description{
Autoregressive time series of order p
}
\examples{
model = getAR(2)
}

