% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{getARMA}
\alias{ARMApq}
\alias{getARMA}
\title{ARMA time series of order p,q}
\usage{
getARMA(p, q)
}
\arguments{
\item{p}{numeric sets the order of the AR part}

\item{q}{numeric sets the order of the MA part}
}
\description{
Returns two functions \code{estimate(x)} und \code{simulate(param,n)} and a string
that contains the name of the model and its parameters.
}
\details{
\enumerate{
\item \code{estimate(x)} estimates the parameters of the \code{ARMA(p,q)} Model by using
the \code{\link[stats]{arima}} function
}
}

