% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QPBoot-class.R
\docType{methods}
\name{getCIs-QPBoot}
\alias{getCIs,QPBoot-method}
\alias{getCIs-QPBoot}
\title{Computes the \eqn{\alpha} an \eqn{1-\alpha} quantiles for the simulated Periodograms
and returns those.}
\usage{
\S4method{getCIs}{QPBoot}(object, alpha = 0.05, method = c("quantiles",
  "norm"))
}
\arguments{
\item{object}{}

\item{alpha}{}
}
\description{
Computes the \eqn{\alpha} an \eqn{1-\alpha} quantiles for the simulated Periodograms
and returns those.
}
\keyword{Access-functions}

