% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{getMA}
\alias{MAq}
\alias{getMA}
\title{Moving-Averages time series of order q}
\usage{
getMA(q)
}
\arguments{
\item{q}{numeric sets the order of the MA(q) process}
}
\description{
Moving-Averages time series of order q
}
\examples{
model = getMA(2)
}

