% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Main.R
\name{main}
\alias{main}
\title{Main function, which does 3 things
\enumerate{
 \item Estimates a parametric \code{model} from a given set of \code{data}
 \item Simulates from that \code{model} and computes the smoothed Quantile
       Periodogram (\link[quantspec]{smoothedPG}) for each simulated time
       series and the given \code{data}
 \item
}}
\usage{
main(data, alpha = 0.05, model = getAR(2), levels = c(0.25, 0.5, 0.75),
  weight = kernelWeight(bw = 0.025), SimNum = 100)
}
\arguments{
\item{data}{}

\item{alpha}{}

\item{model}{}

\item{levels}{}

\item{weight}{}

\item{SimNum}{}
}
\description{
Main function, which does 3 things
\enumerate{
 \item Estimates a parametric \code{model} from a given set of \code{data}
 \item Simulates from that \code{model} and computes the smoothed Quantile
       Periodogram (\link[quantspec]{smoothedPG}) for each simulated time
       series and the given \code{data}
 \item
}
}
\examples{
model = getAR(2)
}

