% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{Models}
\alias{AR}
\alias{ARMA}
\alias{EGARCH}
\alias{GARCH}
\alias{MA}
\alias{Models}
\alias{getAR}
\alias{getARMA}
\alias{getEGARCH}
\alias{getGARCH}
\alias{getMA}
\title{Predefined Time-Series Models}
\usage{
getARMA(spec = list(ar.order = NA, ma.order = NA))

getAR(spec = list(ar.order = 1))

getMA(spec = list(ma.order = 1))

getGARCH(spec = list(alpha = 1, beta = 1))

getEGARCH(spec = list(alpha = 1, beta = 1))
}
\arguments{
\item{spec}{a list specifying the structure of the parameters of the model}
}
\description{
Creates a \link{tsModel-class} object representing a time-series model
}

