% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QPBoot-class.R
\docType{class}
\name{QPBoot-class}
\alias{QPBoot}
\alias{QPBoot-class}
\title{Class for a Parametric Bootstrap based on Quantile Spectral Analysis}
\description{
\code{QPBoot} is a class to compute und contain the results of a Parametric
Bootstrap based on Quantile Spectral Analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{the original data where the parametric bootstrap is based upon}

\item{\code{sPG}}{a smoothed quantile Periodogram from the \code{quantspec}-package
calculatet from \code{data}}

\item{\code{model}}{parametric model for the bootstrap, this is a function that returns a
simulate function and an estimate function see \link{models}}

\item{\code{param}}{parameter estimated for the \code{model} from the \code{data}}

\item{\code{sPGsim}}{smoothed quantile periodograms of the simulated time series}
}}
\keyword{S4-classes}

