% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QPBoot-class.R
\name{QPBoot-constructor}
\alias{QPBoot-constructor}
\alias{qpBoot}
\title{Create an instance of the \code{QPBoot} class by doing 3 things
\enumerate{
 \item Estimates a parametric \code{model} from a given set of \code{data}
 \item Simulates from that \code{model} and computes the smoothed Quantile
       Periodogram (\link[quantspec]{smoothedPG}) for each simulated time
       series and the given \code{data}
 \item returns an object of the class \link{QPBoot} with the calculated
       smoothed Periodograms
}}
\usage{
qpBoot(data, model = getAR(2), levels = c(0.25, 0.5, 0.75),
  weight = kernelWeight(bw = 0.1 * length(data)^(-1/5)), SimNum = 1000,
  trueparam = NULL)
}
\arguments{
\item{data}{}

\item{model}{}

\item{levels}{}

\item{weight}{}

\item{SimNum}{}
}
\description{
Create an instance of the \code{QPBoot} class by doing 3 things
\enumerate{
 \item Estimates a parametric \code{model} from a given set of \code{data}
 \item Simulates from that \code{model} and computes the smoothed Quantile
       Periodogram (\link[quantspec]{smoothedPG}) for each simulated time
       series and the given \code{data}
 \item returns an object of the class \link{QPBoot} with the calculated
       smoothed Periodograms
}
}
\keyword{Constructors}

