% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPBoot-package.R
\docType{package}
\name{QPBoot-package}
\alias{QPBoot}
\alias{QPBoot-package}
\title{Quantile Spectral Analysis for Parametric Bootstrap}
\description{
Methods to validate model assumptions by using Quantile Spectral Analysis and
Paramtric Bootstraps
}
\author{
Stefan Birr
}

