% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsModel-class.R
\docType{methods}
\name{setEstimate-tsModel}
\alias{setEstimate,tsModel-method}
\alias{setEstimate-tsModel}
\title{Sets the estimation Method of a \link{tsModel-class}}
\usage{
\S4method{setEstimate}{tsModel}(object, estimate)
}
\arguments{
\item{object}{the \link{tsModel-class} for that the estimation method shall be defined}

\item{estimate}{a function that has exactly two argumens: data (numeric vector) and spec (a list)
and returns a list containing the estimated parameters.}
}
\value{
Sets the \code{est_function} slot of the \link{tsModel-class}.
}
\description{
Defines the method that should be used when calling \link{Estimate} on the \link{tsModel-class}.
The \code{Estimate} function must have exactly the following two arguments: \code{data} (numeric vector) and 
\code{spec} (a list) and return a list containing the estimated parameters.
}

