% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsModel-class.R
\docType{methods}
\name{setParameter-tsModel}
\alias{setParameter-tsModel}
\alias{setParameter,tsModel-method}
\title{Sets the Parameter of a \link{tsModel-class} manually}
\usage{
\S4method{setParameter}{tsModel}(object, par)
}
\arguments{
\item{object}{the \link{tsModel-class} for that the Parameter will be set}

\item{par}{a list that contains the names and values of the parameters}
}
\value{
Nothing, it sets the \code{par} slot of the \link{tsModel-class}.
}
\description{
This can be used to set the \code{par} slot of a \link{tsModel-class} by hand, in contrast
to call \link{Estimate} on a given data set. The parameter should be a list with the
name an the values of the parameters to be set. See the example below.
}
