% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsModel-class.R
\docType{methods}
\name{setSimulate-tsModel}
\alias{setSimulate,tsModel-method}
\alias{setSimulate-tsModel}
\title{Sets the simulation Method of a \link{tsModel-class}}
\usage{
\S4method{setSimulate}{tsModel}(object, Simulate)
}
\arguments{
\item{object}{the \link{tsModel-class} for that the estimation method shall be defined}

\item{Simulate}{a function that has exactly three argumens: \code{n} (numeric),
\code{spec} (a list) and \code{par} (another list).}
}
\value{
Nothing, it sets the \code{sim_function} slot of the \link{tsModel-class}.
}
\description{
Defines the method that will be used when calling \link{Simulate} on the \link{tsModel-class}.
The passed \code{Simulate} function must have exactly the following three arguments: \code{n} (numeric),
\code{spec} (a list) and \code{par} (another list). It returns a numeric vector with the 
simulated data.
}

