% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsModel-class.R
\docType{methods}
\name{Simulate-tsModel}
\alias{Simulate-tsModel}
\alias{Simulate,tsModel-method}
\title{Simulates from a \link{tsModel-class}}
\usage{
\S4method{Simulate}{tsModel}(object, n)
}
\arguments{
\item{object}{the simulation will be based on this \link{tsModel-class}}

\item{n}{length of the simulated data, will be rounded down if it is not a whole number}
}
\value{
Returns a numeric vector of length \code{n}, that contains data, simulated according to the slot 
\code{sim_function} and \code{par} in the \link{tsModel-class} object.
}
\description{
\code{Simulate} produces a numeric vector of length \code{n} with data simulated according to
the \code{sim_function} defined in the \link{tsModel-class}. To use this function the \code{par} slot
has to be set, either by calling \link{Estimate} or directly via \link{setParameter}.
}
