% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-accessors.R
\name{generics-accessors}
\alias{Estimate}
\alias{Simulate}
\alias{generics-accessors}
\alias{setEstimate}
\alias{setParameter}
\alias{setSimulate}
\title{Generic functions for accessing attributes of objects

These generic functions are needed to access or set the objects' attributes.}
\usage{
Estimate(object, ...)

Simulate(object, ...)

setEstimate(object, ...)

setSimulate(object, ...)

setParameter(object, ...)
}
\arguments{
\item{object}{object from which to get the value}

\item{...}{optional parameters; for documentation see the documentation of
the methods to each of the generic.}
}
\description{
Defines generic function names for the package QPBoot
}

