% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{Models}
\alias{AR}
\alias{ARMA}
\alias{ARMA_GARCH}
\alias{EGARCH}
\alias{GARCH}
\alias{MA}
\alias{Models}
\alias{Noise}
\alias{fGARCH}
\alias{getAR}
\alias{getARCH}
\alias{getARMA}
\alias{getARMA_GARCH}
\alias{getEGARCH}
\alias{getGARCH}
\alias{getMA}
\alias{getNoise}
\title{Predefined Time-Series Models}
\usage{
getARMA(spec = list(ar.order = NA, ma.order = NA))

getAR(spec = list(ar.order = 1))

getMA(spec = list(ma.order = 1))

getNoise()

getGARCH(spec = list(alpha = 1, beta = 1))

getARCH(spec = list(alpha = 1))

getEGARCH(spec = list(alpha = 1, beta = 1))

getARMA_GARCH(spec = list(ar = 1, ma = 1, alpha = 1, beta = 1))
}
\arguments{
\item{spec}{a list specifying the structure of the parameters of the model}
}
\description{
Creates a \link{tsModel-class} object representing a time-series model
}

