% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPBoot-class.R
\docType{class}
\name{QPBoot-class}
\alias{QPBoot-class}
\title{Class for a Parametric Bootstrap based on Quantile Spectral Analysis}
\description{
\code{QPBoot} is a class to compute und contain the results of a parametric
bootstrap based on Quantile Spectral Analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{the original data where the parametric bootstrap is based upon}

\item{\code{sPG}}{a smoothed quantile Periodogram from the \code{quantspec}-package
calculatet from \code{data}}

\item{\code{model}}{parametric model for the bootstrap from \link{tsModel-class}, some Examples
can be found in  \link{Models}}

\item{\code{param}}{parameter estimated for the \code{model} from the \code{data}}

\item{\code{sPGsim}}{smoothed quantile periodograms of the simulated time series}
}}
\keyword{S4-classes}

